requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
ollama